refactor: moved APIs & routes into versioned folder & renamed 'apis' -> 'api'
