+++ backend/routes/v1/index.ts
@@ -2,8 +2,8 @@ import express from 'express';
 
 const router = express.Router();
 
-import controllers from '../controllers/index';
-import middlewares from '../middlewares';
+import controllers from '../../controllers/index';
+import middlewares from '../../middlewares';
 
 /**
  * @openapi
